requirements.txt
fastapi
uvicorn
requests
python-dotenv
setuptools
pyrebase4
firebase_admin